+++ bytesmart-admin-center/src/main/java/com/bytesmart/admincenter/domain/BytesmartEmployeeRole.java
@@ -4,22 +4,22 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
 
 public class BytesmartEmployeeRole {
-    private Integer employeeId;
-    private Integer roleId;
+    private Long employeeId;
+    private Long roleId;
 
-    public Integer getEmployeeId() {
+    public Long getEmployeeId() {
         return employeeId;
     }
 
-    public void setEmployeeId(Integer employeeId) {
+    public void setEmployeeId(Long employeeId) {
         this.employeeId = employeeId;
     }
 
-    public Integer getRoleId() {
+    public Long getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Integer roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId = roleId;
     }
     @Override
