+++ bytesmart-admin-center/src/main/java/com/bytesmart/admincenter/mapper/BytesmartPostMapper.java
@@ -10,7 +10,7 @@ public interface BytesmartPostMapper {
     public List<BytesmartPost> selectPostList(BytesmartPost post);
 
     //根据id查询
-    public BytesmartPost selectPostById(Integer postId);
+    public BytesmartPost selectPostById(Long postId);
 
     //修改
     public int updatePost(BytesmartPost post);
@@ -24,7 +24,7 @@ public interface BytesmartPostMapper {
     public List<BytesmartPost> selectPostAll();
 
     //通过用户id获取岗位选择框列表
-    public List<Integer> selectPostListByEmployeeId(Integer employeeId);
+    public List<Integer> selectPostListByEmployeeId(Long employeeId);
 
     //通过用户名查所属岗位
     public List<BytesmartPost> selectPostListByUserName(String username);
